Tidy up scratchstack-shapegen
Removed unused code.
Made tests not write to stdout.
